+++ ext/src/utils.jl
@@ -79,6 +79,7 @@ record_interaction(file, figure; kwargs...)
 
 """
     subscript(i::Int)
+
 Transform `i` to a string that has `i` as a subscript.
 """
 function subscript(i::Int)
@@ -111,6 +112,7 @@ end
 
 """
     superscript(i::Int)
+
 Transform `i` to a string that has `i` as a superscript.
 """
 function superscript(i::Int)
@@ -161,7 +163,7 @@ end
     to_alpha(c, α = 0.75)
 Create a color same as `c` but with given alpha channel.
 """
-function to_alpha(c, α = 1.2)
+function to_alpha(c, α = 0.75)
     c = to_color(c)
     return RGBAf(c.r, c.g, c.b, α)
 end